+++ scripts/prerequisites/helper_scripts/generate/sql/oracle/createOS1DB.sql
@@ -1,30 +1,43 @@
--- Please ensure you already have existing oracle instance or pluggable database (PDB). If not, please create one first
+-- *********************************************************************
+-- IBM FileNet Content Manager ObjectStore preparation script for Oracle
+-- *********************************************************************
+-- Usage:
+-- Use Oracle sql command-line to execute the template file using @{file} option and user as SYSDBA
+-- sqlplus / as sysdba
+-- @createOS1DB.sql
 
--- create tablespace
+-- Please ensure you already have existing oracle instance.
+-- If your oracle instance does not support multi-tenant architecture, comment out follow lines:
+CREATE PLUGGABLE DATABASE ${os_name} ADMIN USER ${os_name}_admin IDENTIFIED BY ${yourpassword} ROLES=(DBA);
+ALTER PLUGGABLE DATABASE ${os_name} OPEN READ WRITE;
+ALTER PLUGGABLE DATABASE ${os_name} save state;
+ALTER SESSION SET CONTAINER=${os_name};
+
+-- Create tablespace
 -- Change DATAFILE/TEMPFILE as required by your configuration
 CREATE TABLESPACE ${os_name}DATATS DATAFILE '/home/oracle/orcl/${os_name}DATATS.dbf' SIZE 200M REUSE AUTOEXTEND ON NEXT 20M EXTENT MANAGEMENT LOCAL SEGMENT SPACE MANAGEMENT AUTO ONLINE PERMANENT;
 CREATE TEMPORARY TABLESPACE ${os_name}DATATSTEMP TEMPFILE '/home/oracle/orcl/${os_name}DATATSTEMP.dbf' SIZE 200M REUSE AUTOEXTEND ON NEXT 20M EXTENT MANAGEMENT LOCAL;
 
--- create a new user for ${youruser1}
+-- Create a new user for ${youruser1}
 CREATE USER ${youruser1} PROFILE DEFAULT IDENTIFIED BY ${yourpassword} DEFAULT TABLESPACE ${os_name}DATATS TEMPORARY TABLESPACE ${os_name}DATATSTEMP ACCOUNT UNLOCK;
 
--- provide quota on all tablespaces with BPM tables
+-- Provide quota on all tablespaces with BPM tables
 ALTER USER ${youruser1} QUOTA UNLIMITED ON ${os_name}DATATS;
 ALTER USER ${youruser1} DEFAULT TABLESPACE ${os_name}DATATS;
 ALTER USER ${youruser1} TEMPORARY TABLESPACE ${os_name}DATATSTEMP;
 
--- allow the user to connect to the database
+-- Allow the user to connect to the database
 GRANT CONNECT TO ${youruser1};
 GRANT ALTER session TO ${youruser1};
 
--- grant privileges to create database objects
+-- Grant privileges to create database objects
 GRANT CREATE SESSION TO ${youruser1};
 GRANT CREATE TABLE TO ${youruser1};
 GRANT CREATE VIEW TO ${youruser1};
 GRANT CREATE SEQUENCE TO ${youruser1};
 GRANT CREATE PROCEDURE TO ${youruser1};
 
--- grant access rights to resolve XA related issues
+-- Grant access rights to resolve XA related issues
 GRANT SELECT on pending_trans$ TO ${youruser1};
 GRANT SELECT on dba_2pc_pending TO ${youruser1};
 GRANT SELECT on dba_pending_transactions TO ${youruser1};