+++ scripts/prerequisites/helper_scripts/generate/sql/sqlserver/createOS1DB.sql
@@ -1,3 +1,12 @@
+-- ************************************************************************
+-- IBM FileNet Content Manager ObjectStore preparation script for SQLServer
+-- ************************************************************************
+-- Usage:
+-- Use sqlcmd command-line tool to execute the template file using -i option and
+-- user with privileges to create databases and filegroups
+-- sqlcmd -S serverName\instanceName -U dbaUser -P dbaPassword -i C:\createOS1DB.sql
+
+
 -- create ${os_name} object store database, you could update FILENAME as your requirement.
 -- Please make sure you change the drive and path to your MSSQL database.
 CREATE DATABASE ${os_name}